Indexing at startup, not at first search
(cherry picked from commit f8d5aff0d5f262769f20d17186918d1ba0e415b9)
