+++ lib/utils/constants.dart
@@ -13,3 +13,4 @@ enum CryptMode { encrypt, decrypt }
 const UNKNOWN_ELEMENT = '<?>';
 
 enum AlphabetModificationMode { J_TO_I, C_TO_K, W_TO_VV, REMOVE_Q }
+
